Initial draft of new folder structure with new changes to import lines, and finding files
