--- currency_conversions.csv
@@ -1,65 +1,0 @@
-,currency_name,conversion_rate,conversion_rate_percentage
-0,,,
-1,US Dollar,1.270210,0.787271
-2,Euro,1.159420,0.862500
-3,Indian Rupee,104.236815,0.009594
-4,Australian Dollar,1.913547,0.522590
-5,Canadian Dollar,1.679755,0.595325
-6,Singapore Dollar,1.716754,0.582495
-7,Swiss Franc,1.137497,0.879123
-8,Malaysian Ringgit,5.934040,0.168519
-9,Japanese Yen,182.973924,0.005465
-10,Chinese Yuan Renminbi,9.200967,0.108684
-11,,,
-12,Argentine Peso,324.156609,0.003085
-13,Australian Dollar,1.913547,0.522590
-14,Bahraini Dinar,0.477599,2.093806
-15,Botswana Pula,17.017506,0.058763
-16,Brazilian Real,6.111102,0.163637
-17,Bruneian Dollar,1.716754,0.582495
-18,Bulgarian Lev,2.267628,0.440989
-19,Canadian Dollar,1.679755,0.595325
-20,Chilean Peso,1013.340164,0.000987
-21,Chinese Yuan Renminbi,9.200967,0.108684
-22,Colombian Peso,5292.853093,0.000189
-23,Czech Koruna,27.441560,0.036441
-24,Danish Krone,8.634876,0.115809
-25,Emirati Dirham,4.664847,0.214369
-26,Euro,1.159420,0.862500
-27,Hong Kong Dollar,9.948770,0.100515
-28,Hungarian Forint,429.907717,0.002326
-29,Icelandic Krona,172.202772,0.005807
-30,Indian Rupee,104.236815,0.009594
-31,Indonesian Rupiah,19042.384456,0.000053
-32,Iranian Rial,53746.518725,0.000019
-33,Israeli Shekel,4.664675,0.214377
-34,Japanese Yen,182.973924,0.005465
-35,Kazakhstani Tenge,573.284281,0.001744
-36,Kuwaiti Dinar,0.390310,2.562067
-37,Libyan Dinar,6.105676,0.163782
-38,Malaysian Ringgit,5.934040,0.168519
-39,Mauritian Rupee,58.002953,0.017241
-40,Mexican Peso,21.657187,0.046174
-41,Nepalese Rupee,166.857082,0.005993
-42,New Zealand Dollar,2.083844,0.479882
-43,Norwegian Krone,13.666104,0.073174
-44,Omani Rial,0.489024,2.044888
-45,Pakistani Rupee,361.755848,0.002764
-46,Philippine Peso,70.251495,0.014235
-47,Polish Zloty,5.177865,0.193130
-48,Qatari Riyal,4.623565,0.216283
-49,Romanian New Leu,5.749322,0.173934
-50,Russian Ruble,108.870362,0.009185
-51,Saudi Arabian Riyal,4.763288,0.209939
-52,Singapore Dollar,1.716754,0.582495
-53,South African Rand,23.634671,0.042311
-54,South Korean Won,1662.214320,0.000602
-55,Sri Lankan Rupee,390.781480,0.002559
-56,Swedish Krona,13.626611,0.073386
-57,Swiss Franc,1.137497,0.879123
-58,Taiwan New Dollar,39.410295,0.025374
-59,Thai Baht,45.256182,0.022096
-60,Trinidadian Dollar,8.620822,0.115998
-61,Turkish Lira,33.107531,0.030205
-62,US Dollar,1.270210,0.787271
-63,Venezuelan Bolivar,3509724.742459,0.000000
