+++ source_data_files/currency_conversions.csv
@@ -1,0 +1,65 @@
+,currency_name,conversion_rate,conversion_rate_percentage
+0,,,
+1,US Dollar,1.263541,0.791426
+2,Euro,1.158866,0.862912
+3,Indian Rupee,103.642111,0.009649
+4,Australian Dollar,1.908948,0.523849
+5,Canadian Dollar,1.674117,0.597330
+6,Singapore Dollar,1.709122,0.585096
+7,Swiss Franc,1.134148,0.881719
+8,Malaysian Ringgit,5.901706,0.169443
+9,Japanese Yen,182.250801,0.005487
+10,Chinese Yuan Renminbi,9.155847,0.109220
+11,,,
+12,Argentine Peso,323.243985,0.003094
+13,Australian Dollar,1.908948,0.523849
+14,Bahraini Dinar,0.475092,2.104858
+15,Botswana Pula,16.933232,0.059055
+16,Brazilian Real,6.118627,0.163435
+17,Bruneian Dollar,1.709122,0.585096
+18,Bulgarian Lev,2.266545,0.441200
+19,Canadian Dollar,1.674117,0.597330
+20,Chilean Peso,1010.813706,0.000989
+21,Chinese Yuan Renminbi,9.155847,0.109220
+22,Colombian Peso,5257.146126,0.000190
+23,Czech Koruna,27.462851,0.036413
+24,Danish Krone,8.630506,0.115868
+25,Emirati Dirham,4.640355,0.215501
+26,Euro,1.158866,0.862912
+27,Hong Kong Dollar,9.897180,0.101039
+28,Hungarian Forint,430.488008,0.002323
+29,Icelandic Krona,172.084644,0.005811
+30,Indian Rupee,103.642111,0.009649
+31,Indonesian Rupiah,18938.448851,0.000053
+32,Iranian Rial,53496.774315,0.000019
+33,Israeli Shekel,4.642159,0.215417
+34,Japanese Yen,182.250801,0.005487
+35,Kazakhstani Tenge,569.444463,0.001756
+36,Kuwaiti Dinar,0.388496,2.574029
+37,Libyan Dinar,6.076296,0.164574
+38,Malaysian Ringgit,5.901706,0.169443
+39,Mauritian Rupee,58.174114,0.017190
+40,Mexican Peso,21.580543,0.046338
+41,Nepalese Rupee,165.905109,0.006028
+42,New Zealand Dollar,2.076103,0.481672
+43,Norwegian Krone,13.622484,0.073408
+44,Omani Rial,0.486466,2.055643
+45,Pakistani Rupee,361.987035,0.002763
+46,Philippine Peso,69.843914,0.014318
+47,Polish Zloty,5.190491,0.192660
+48,Qatari Riyal,4.599290,0.217425
+49,Romanian New Leu,5.751241,0.173876
+50,Russian Ruble,109.193668,0.009158
+51,Saudi Arabian Riyal,4.738280,0.211047
+52,Singapore Dollar,1.709122,0.585096
+53,South African Rand,23.604020,0.042366
+54,South Korean Won,1654.988507,0.000604
+55,Sri Lankan Rupee,388.109621,0.002577
+56,Swedish Krona,13.651343,0.073253
+57,Swiss Franc,1.134148,0.881719
+58,Taiwan New Dollar,39.215678,0.025500
+59,Thai Baht,44.923451,0.022260
+60,Trinidadian Dollar,8.575453,0.116612
+61,Turkish Lira,32.889838,0.030405
+62,US Dollar,1.263541,0.791426
+63,Venezuelan Bolivar,3491747.264056,0.000000
