+++ source_data_files/currency_conversions_test.csv
@@ -1,0 +1,65 @@
+,currency_name,conversion_rate,conversion_rate_percentage
+0,,,
+1,US Dollar,1.264937,0.790553
+2,Euro,1.158331,0.863311
+3,Indian Rupee,103.773338,0.009636
+4,Australian Dollar,1.912939,0.522756
+5,Canadian Dollar,1.675713,0.596761
+6,Singapore Dollar,1.710671,0.584566
+7,Swiss Franc,1.133626,0.882125
+8,Malaysian Ringgit,5.908513,0.169247
+9,Japanese Yen,182.523274,0.005479
+10,Chinese Yuan Renminbi,9.165191,0.109108
+11,,,
+12,Argentine Peso,323.667306,0.003090
+13,Australian Dollar,1.912939,0.522756
+14,Bahraini Dinar,0.475616,2.102535
+15,Botswana Pula,16.955287,0.058979
+16,Brazilian Real,6.122256,0.163338
+17,Bruneian Dollar,1.710671,0.584566
+18,Bulgarian Lev,2.265498,0.441404
+19,Canadian Dollar,1.675713,0.596761
+20,Chilean Peso,1012.347658,0.000988
+21,Chinese Yuan Renminbi,9.165191,0.109108
+22,Colombian Peso,5244.457956,0.000191
+23,Czech Koruna,27.430666,0.036456
+24,Danish Krone,8.627639,0.115907
+25,Emirati Dirham,4.645481,0.215263
+26,Euro,1.158331,0.863311
+27,Hong Kong Dollar,9.908189,0.100927
+28,Hungarian Forint,429.996208,0.002326
+29,Icelandic Krona,172.019644,0.005813
+30,Indian Rupee,103.773338,0.009636
+31,Indonesian Rupiah,18956.778080,0.000053
+32,Iranian Rial,53603.056380,0.000019
+33,Israeli Shekel,4.646574,0.215212
+34,Japanese Yen,182.523274,0.005479
+35,Kazakhstani Tenge,570.477392,0.001753
+36,Kuwaiti Dinar,0.388846,2.571709
+37,Libyan Dinar,6.071861,0.164694
+38,Malaysian Ringgit,5.908513,0.169247
+39,Mauritian Rupee,58.150240,0.017197
+40,Mexican Peso,21.569093,0.046363
+41,Nepalese Rupee,166.115172,0.006020
+42,New Zealand Dollar,2.078234,0.481178
+43,Norwegian Krone,13.602056,0.073518
+44,Omani Rial,0.487009,2.053352
+45,Pakistani Rupee,362.453283,0.002759
+46,Philippine Peso,69.914010,0.014303
+47,Polish Zloty,5.186679,0.192802
+48,Qatari Riyal,4.604370,0.217185
+49,Romanian New Leu,5.748936,0.173945
+50,Russian Ruble,108.884815,0.009184
+51,Saudi Arabian Riyal,4.743514,0.210814
+52,Singapore Dollar,1.710671,0.584566
+53,South African Rand,23.604332,0.042365
+54,South Korean Won,1654.505739,0.000604
+55,Sri Lankan Rupee,388.582903,0.002573
+56,Swedish Krona,13.623404,0.073403
+57,Swiss Franc,1.133626,0.882125
+58,Taiwan New Dollar,39.231705,0.025490
+59,Thai Baht,44.981195,0.022232
+60,Trinidadian Dollar,8.583557,0.116502
+61,Turkish Lira,32.947847,0.030351
+62,US Dollar,1.264937,0.790553
+63,Venezuelan Bolivar,3501511.055274,0.000000
